Convert React Kanban app to TypeScript
This commit converts the React-based Kanban application from JavaScript to TypeScript.

Key changes include:
- Added TypeScript and relevant @types packages to devDependencies.
- Configured TypeScript with a tsconfig.json file.
- Renamed .js/.jsx files to .ts/.tsx.
- Defined interfaces and types for component props (Task, ColumnProps, TaskCardProps), state (BoardState with Task[]), and data models (TaskStatus).
- Updated all components (App, Board, Column, TaskCard, index) and event handlers to use TypeScript types.
- Adjusted Jest tests (Board.test.tsx) for TypeScript compatibility.

Note: Due to sandbox environment limitations, I couldn't fully complete the package installation, so direct execution of tsc for type checking and npm test for Jest tests could not be performed. The conversion is based on static code changes and adherence to TypeScript best practices.
