+++ kanban-react-app/package.json
@@ -24,6 +24,13 @@
       "react-app/jest"
     ]
   },
+  "devDependencies": {
+    "typescript": "^5.0.0",
+    "@types/node": "^20.0.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/jest": "^29.0.0"
+  },
   "browserslist": {
     "production": [
       ">0.2%",
