+++ kanban-react-app/src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import Board from './Board';
 
-function App() {
+const App: React.FC = () => { // Explicitly typed as React.FC
   return (
     <div className="App">
       <Board />
